xboxrt: Move supplemental C functions, add own Makefile
