+++ lib/hal/debug.c
