+++ lib/xboxrt/Makefile
@@ -1,0 +1,4 @@
+SRCS += \
+	$(NXDK_DIR)/lib/xboxrt/libc_extensions/stat.c \
+	$(NXDK_DIR)/lib/xboxrt/libc_extensions/strings.c \
+	$(NXDK_DIR)/lib/xboxrt/libc_extensions/wchar.c
