+++ lib/xboxrt/libc_extensions/stat.c
