+++ lib/xboxrt/libc_extensions/stat.h
