+++ lib/xboxrt/libc_extensions/strings.c
