+++ lib/xboxrt/libc_extensions/strings.h
