+++ lib/xboxrt/libc_extensions/wchar.c
